requirements.txt
asgiref==3.5.0
configparser==5.2.0
Django==4.0.4
django-phonenumber-field==6.1.0
django-taggit==2.1.0
importlib-metadata==4.11.3
Markdown==3.3.6
mysqlclient==2.1.0
phonenumbers==8.12.46
Pillow==9.1.0
python-slugify==6.1.1
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2022.1
zipp==3.8.0
